concludes unfinished merge conflicts and updates tests for pipeline.
